requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
pypi==2.1
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyserial==3.5
websocket-client==1.2.1
websockets==10.0
ws4py==0.5.1
